pyproject.toml
[project]
name = "myst-kinotate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedgen",
    "ipython!=8.7.0",
    "jupyterlab-myst",
    "mystmd",
    "pyyaml",
    "rich",
]
